+++ src/cmd/go/internal/cfg/bench_test.go
@@ -5,6 +5,7 @@
 package cfg
 
 import (
+	"cmd/internal/pathcache"
 	"internal/testenv"
 	"testing"
 )
@@ -13,7 +14,7 @@ func BenchmarkLookPath(b *testing.B) {
 	testenv.MustHaveExecPath(b, "go")
 	b.ResetTimer()
 	for i := 0; i < b.N; i++ {
-		_, err := LookPath("go")
+		_, err := pathcache.LookPath("go")
 		if err != nil {
 			b.Fatal(err)
 		}
